pyproject.toml
[project]
name = "fastapi-docker"
version = "0.1.0"
description = ""
authors = [
    {name = "Fabio Lucas",email = "fabiolucamz@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard] (>=0.135.1,<0.136.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "sqlalchemy (>=2.0.48,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "psycopg2 (>=2.9.11,<3.0.0)",
    "bcrypt (>=5.0.0,<6.0.0)",
    "python-jose (>=3.5.0,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
